requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# Vector store
chromadb==0.4.22

# PDF processing
pypdf==3.17.4

# OpenAI
openai==1.10.0

# Environment variables
python-dotenv==1.0.0

# Text splitting
tiktoken==0.5.2

# Additional utilities
numpy==1.26.3
